Add files via upload
There was a missing '{' on line 44 and a subsequent missing '}' on line 46.
I changed the styling to a four space indent due the initial styling being a mix of four space and eight space indents. Additional to the styling changes, I added spaces to variable and brace locations to make it easier to read.
Finally, I added a "default: break;" to lines 66 and 67.